move mus to new .explain, keep imports for backwards compatibility
